chor: add global exception handler and response entity

+++ Task Management Service/src/main/java/com/parkconnect/task_management/service/ProjectServiceImpl.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.parkconnect.task_management.dto.ProjectDetailDto;
 import com.parkconnect.task_management.dto.ProjectDto;
-import com.parkconnect.task_management.entity.Project;
+import com.parkconnect.task_management.entity.ProjectEntity;
+import com.parkconnect.task_management.exception.ResourceNotFoundException;
 import com.parkconnect.task_management.mapper.ProjectMapper;
 import com.parkconnect.task_management.repository.ProjectRepository;
 
@@ -19,10 +19,11 @@ public class ProjectServiceImpl implements ProjectService {
         this.projectListRepository = projectListRepository;
     }
 
-    public Project addNewProject(ProjectDetailDto project) {
-
-        Project newProject  = ProjectMapper.ProjectDetailDtoToProjectEntity(project);
-        return projectListRepository.save(newProject);
+    public ProjectDto addNewProject(ProjectDto projectDto) {
+        ProjectEntity projectEntity = new ProjectEntity();
+        ProjectEntity newProject  = ProjectMapper.ProjectDetailDtoToProjectEntity(projectDto, projectEntity);
+        ProjectEntity result = projectListRepository.save(newProject);
+        return ProjectMapper.ProjectEntityToProjectDetailDto(result);
     }
 
     public List<ProjectDto> getProjects() {
@@ -33,24 +34,28 @@ public class ProjectServiceImpl implements ProjectService {
         return projectList;
     }
 
-    public ProjectDetailDto getProjectById(Integer projectId) {
-        Project project =  projectListRepository.findById(projectId).orElse(null);
-        if (project != null) {
-            return new ProjectDetailDto(project.getProjectId(), project.getTitle(), project.getDescription(), project.getStartDate());
-        }
-        return null;
+    public ProjectDto getProjectById(Integer projectId) {
+        ProjectEntity project =  projectListRepository.findById(projectId).orElseThrow(() -> new ResourceNotFoundException("Project Not Found"));
+    
+        return new ProjectDto(project.getProjectId(), project.getTitle(), project.getDescription(), project.getStartDate());
     }
 
     public void deleteProject(Integer projectId) {
+        Boolean isProjectIdExist =  projectListRepository.existsById(projectId);
+    
+        if(!isProjectIdExist){
+            throw new ResourceNotFoundException("Project with id " + projectId +" is Not Found");
+        }
         projectListRepository.deleteById(projectId);
     }
 
-    public String updateProject(Integer projectId, ProjectDetailDto projectDetailDto) {
-        // Project project = projectListRepository.findById(projectId).orElse(null);
+    public String updateProject(Integer projectId, ProjectDto projectDetailDto) {
+        ProjectEntity existingProject = projectListRepository.findById(projectId).orElseThrow(() -> new ResourceNotFoundException("Project Not Found"));
+
         projectDetailDto.setProjectId(projectId);
-        Project filteredProjectDetail = ProjectMapper.ProjectDetailDtoToProjectEntity(projectDetailDto);
-        // filteredProjectDetail.setProjectId(projectId);
+        ProjectEntity filteredProjectDetail = ProjectMapper.ProjectDetailDtoToProjectEntity(projectDetailDto, existingProject);
         projectListRepository.save(filteredProjectDetail);
+
         return "project added successfully";
     }
 }
